+++ lib/src/models.dart
@@ -86,6 +86,9 @@ enum MapType {
 
   /// MapyCZ
   mapyCz,
+
+  // MAPPLS MapmyIndia
+  mappls,
 }
 
 /// Defines the supported modes of transportation for [showDirections]